feat(operator): add proper health checks
Use dedicated controller-runtime facilities to create health and
readiness checks for the webhook server.

Unfortunately it seems like this doesn't solve the webhook
unavailability issue right after starting the operator. I've added an
explicit 10 second wait to the E2E tests for now, we'll have to see if
that's enough to fix the problem.